fix: move stubs directory
