+++ frontend/contexts/__stubs__/ThemeProviderChild.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { ThemeContext } from '../../ThemeProvider';
+import { ThemeContext } from '../ThemeProvider';
 import { useContext } from 'react';
-import { ThemeMode } from '../../../types/styles';
+import { ThemeMode } from '../../types/styles';
 
 type Props = {
   modeToSet: ThemeMode;
