+++ frontend/tsconfig.json
@@ -16,5 +16,5 @@
     "incremental": true
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
-  "exclude": ["node_modules", ".storybook", "utils-test", "**/__stories__/**", "**/__tests__/**"]
+  "exclude": ["node_modules", ".storybook", "utils-test", "**/__stories__/**", "**/__tests__/**", "**/__stubs__/**"]
 }
